+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.2.12
+- New feature: Automatically generated portal hub
 - Fix NoPortalsPreventsContruction not working [#29](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/29)
 
 ### v0.2.11